requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
beautifulsoup4==4.7.1
Brlapi==0.6.6
bs4==0.0.1
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cssselect2==0.2.1
cupshelpers==1.0
defer==1.0.6
defusedxml==0.5.0
distro-info==0.18
fuzzywuzzy==0.17.0
html5lib==1.0.1
httplib2==0.9.2
idna==2.6
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
meld==3.18.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
peewee==3.7.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
Pyphen==0.9.5
pyRFC3339==1.0
python-apt==1.6.0
python-debian==0.1.32
python-slugify==3.0.2
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
soupsieve==1.9.1
ssh-import-id==5.7
system-service==0.3
systemd-python==234
text-unidecode==1.2
tinycss2==0.6.1
tornado==4.5.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unicode==2.7
Unidecode==1.1.1
urllib3==1.22
usb-creator==0.3.3
vboxapi==1.0
wadllib==1.3.2
WeasyPrint==43
webencodings==0.5.1
xkit==0.0.0
zope.interface==4.3.2
